# backend/routes/spec/api-docs.yaml
openapi: "3.0.0"
info:
  version: 0.0.1
  title: Digital Realty Website Backend API
  description: |
    Provides the backend/bridge for the digital realty website
  contact:
    name: Propane Studios
servers:
  - url: /api/v1
    description: Version 1 of this server
security:
  - basicAuth: []
paths:
  /user:
    get:
      description: Get a list of all users. Logged in user must be in role solutions architect, sales, or admin to receive the user list. Any other role gets denied
      summary: List all users
      operationId: listUsers
      tags:
        - User
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/Limit"
        - name: page
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/Page"
        - name: firstName
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/UserFirstNameField'
          examples:
            exactMatch:
              summary: Match first name exactly
              value: admin
            startsWith:
              summary: Match any name starting with
              value: 'adm%'
            endsWith:
              summary: Match any name ending with
              value: '%min'
            contains:
              summary: Match any name containing
              value: dmi%
        - name: lastName
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/UserLastNameField'
          examples:
            exactMatch:
              summary: Match last name excatly
              value: admin
            startsWith: 
              summary: Match any name starting with
              value: adm%
            endsWith:
              summary: Match any name ending with
              value: '%min'
            contains:
              summary: Match any name containing
              value: dmi%  
        - name: role
          in: query
          description: User's role ID
          required: false
          schema:
            type: string
          example: 1624e83e-fc72-4e81-ad04-3396103b95ad
        - name: email
          in: query
          description: User's email address
          required: false
          schema:
            $ref: "#/components/schemas/EmailFields"
        - name: phone
          in: query
          description: User's phone number
          required: false
          schema:
            $ref: '#/components/schemas/PhoneFields'
        - name: order
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/OrderFields'
        - name: archived
          in: query
          description: User account status (active or archived)
          required: false
          schema:
            $ref: "#/components/schemas/ArchivedFields"
          examples:
            active: 
              summary: Active Account
              value: false
            archived:
              summary: Archived Account
              value: true
      responses:
        '200':
          description: A list of users that the logged in user has access to
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      $ref: "#/components/schemas/User"
                  total:
                    $ref: "#/components/schemas/Total"
                  numPages:
                    $ref: "#/components/schemas/NumPages"
        '400':
          description: Invalid Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvalidRequest"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotAuthorized"
        '500':
            description: Unexpected error
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/Error"
    post:
      description: Create a user. The logged in user must be in the role solutions architect, sales, or admin to create the user account. Any other roles get denied
      operationId: createUser
      summary: Create a user.
      tags:
        - User
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostUser"
      responses:
        '200':
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        '400':
          description: Invalid request
          content:
            application/json:
              schema:  
                $ref: "#/components/schemas/InvalidRequest"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotAuthorized"
        '500':
          description: An error occurred
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /user/{id}:
    get:
      description: Get a user by its ID. The logged in user must be in role solutions architect, sales, or admin to find a user. Any other role gets denied
      summary: Get a user
      tags:
        - User
      parameters:
        - name: id
          in: path
          description: The user UUID
          required: true
          schema:
            $ref: '#/components/schemas/IdFields'
      responses:
        '200':
          description: The requested user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        '400':
          description: Invalid Request
          content:
            application/json:
              schema:  
                $ref: "#/components/schemas/InvalidRequest"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotAuthorized"
        '404':
          description: error not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFound"
        '500':
            description: Unexpected error
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/Error"
    put:
      description: Update a user. The logged in user must be in role solutions architect, sales, or admin to update a user. Any other role gets denied
      operationId: update user
      summary: Update user data
      tags:
        - User
      requestBody:
        required: true
        content: 
          application/json:
            schema:
              $ref: "#/components/schemas/PutUser"
      responses:
        '200':
          description: Successfully updated the user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PutUser"

        '400':
          description: Invalid request
          content:
            application/json:
              schema:  
                $ref: "#/components/schemas/InvalidRequest"
        '500':
          description: An error occurred
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      description: Delete a user. The logged in user must be in role solutions architect, sales, or admin to delete (or archive) a user. Any other role gets denied
      summary: Delete a user
      tags:
        - User
      parameters:
        - name: id
          in: path
          description: To delete user it will need UUID
          required: true
          schema:
            $ref: '#/components/schemas/IdFields'
      responses:
        "200":
          description: Successfully deleted (archived) the user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessUser"
        '500':
          description: An error occurred
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /user/resetpassword:
    post:
      description: Reset user's password
      summary: Reset password
      tags:
        - User
      parameters:
        - name: email
          in: query
          description: The user's email
          required: true
          schema:
            type: string
        - name: password
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/UserPasswordField"
        - name: token
          in: query
          description: The user's token
          required: true
          schema:
            type: string
      responses:
        '202':
          description: success message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResetpassword"
        '404':
          description: error resseting password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFound"
  /user/forgotpassword:
    post:
      description: forgot a user's password
      summary: Forgot password
      tags:
        - User
      parameters:
        - name: email
          in: query
          description: The user's email
          required: false
          schema:
            type: string
      requestBody:
        required: true 
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserErrorForgotPassword"
      responses:
        '404':
          description: error resseting password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFound"
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
  schemas:
    Total:
      type: number
      description: total number of records
    NumPages:
      type: number
      description: total number of pages given the limit
    Limit:
      type: number
      description: Total number of records per page (max 100)
    Page:
      description: Which page to start on (relative to limit)
      type: integer
      format: int32
    PhoneFields:
      type: string
      description: User's phone number
      maxLength: 20
      pattern: '/^\s*(?:\+?(\d{1,3}))?[-. (]*(\d{3})[-. )]*(\d{3})[-. ]*(\d{4})(?: *x(\d+))?\s*$/'
      example: 1-555-555-5555
    EmailFields:
      type: string
      description: User's email address
      pattern: /^\w+([.-]?\w+)+([+]?(\d|\w))*@\w+([.-]?\w+)*(\.\w{2,3})+$/
      example: 'admin@pdxmodeler.com'
    OrderFields:
      description: Fields (i.e. archived) followed by their sort order (ASC/DESC)
      type: array
      items:
        type: array
        items: 
          type: string
      example: [['archived', 'DESC']]
    UserFirstNameField:
      type: string
      description: First Name of the User
    UserLastNameField:
      type: string
      description: Last Name of the User
...
    UsernameField:
      type: string
      maxLength: 60
      description: user's username (for login)
    UserRoleField:
      type: string
      description: User's role ID
      example: b0d0ab14-4069-48df-9258-86647a0d7d95
    UserPasswordField:
      type: string
      maxLength: 60
      description: Password for user
    TokenField:
      type: string
      description: contains error message (error resetting password)
...
    User:
      required:
        - id
        - firstName
        - lastName
        - email
        - phone
        - username
        - role
        - archived
      properties:
        id:
          $ref: '#/components/schemas/IdFields'
        firstName:
          $ref: '#/components/schemas/UserFirstNameField'
        lastName:
          $ref: '#/components/schemas/UserLastNameField'
        email:
          $ref: "#/components/schemas/EmailFields"
        phone:
          $ref: "#/components/schemas/PhoneFields"
        username:
          $ref: '#/components/schemas/UsernameField'
        role:
          $ref: '#/components/schemas/UserRoleField'
        archived:
          $ref: '#/components/schemas/ArchivedFields'
    PostUser:
      required:
        - firstName
        - lastName
        - email
        - phone
        - username
        - role
        - archived
        - password
      properties:
        firstName:
          $ref: '#/components/schemas/UserFirstNameField'
        lastName:
          $ref: '#/components/schemas/UserLastNameField'
        email:
          email:
          $ref: "#/components/schemas/EmailFields"
        phone:
          $ref: "#/components/schemas/PhoneFields"
        username:
          $ref: '#/components/schemas/UsernameField'
        password:
          $ref: "#/components/schemas/UserPasswordField"
        role:
          $ref: '#/components/schemas/UserRoleField'
    PutUser:
      properties:
        firstName:
          $ref: '#/components/schemas/UserFirstNameField'
        lastName:
          $ref: '#/components/schemas/UserLastNameField'
        email:
          email:
          $ref: "#/components/schemas/EmailFields"
        phone:
          $ref: "#/components/schemas/PhoneFields"
        username:
          $ref: '#/components/schemas/UsernameField'
        role:
          $ref: '#/components/schemas/UserRoleField'
    SuccessUser:
      required:
        - user
      properties:
        user:
          type: array
          $ref: "#/components/schemas/User"
          description: Contains the user that was successfully created/updated
...
    InvalidRequest:
      required:
        - error
      properties:
        error:
          type: string
          description: Contains a message detailing why the request was invalid
    Error:
      required:
        - error
      properties:
        error:
          type: string
          description: contains a message about the failure
...
    UserErrorForgotPassword:
      required:
        - error
      properties:
        email:
          $ref: "#/components/schemas/EmailFields"
    SuccessResetpassword:
      required:
        - error
      properties:
        email:
          $ref: "#/components/schemas/EmailFields"
        password:
          $ref: "#/components/schemas/UserPasswordField"
        token:
          $ref: "#/components/schemas/TokenField"
      description: Contains the user password was successfully reset
...
    IdFields:
      type: string
      description: Object id (UUID)
      example: b0d0ab14-4069-48df-9258-86647a0d7d95
...
    ArchivedFields:
      type: boolean
      description: Whether or not this is archived
      example: false
...
    NotFound:
      required:
        - error
      properties:
        error:
          type: string
          example: "Not Found"
    NotAuthorized:
      required:
        - error
      properties:
        error:
          type: string
          example: "Not Authorized"